Use bind variables for types which do not support literals.
JdbcType gets a new method hasLiteralForm to check if a literal
representation is supported. This must return false in cases where
valueToSQLLiteral would otherwise throw a SlickException.

QueryBuilder (and driver-specific subclasses thereof) uses this method
to treat LiteralNodes as volatile (i.e. using bind variables) as needed.

In addition to the types which already threw an exception, Blob and Clob
are marked as not having a literal form in JdbcTypesComponent.

Tests in JdbcTypeTest (all tests now execute with both, volatile-hinted
and not volatile-hinted literals). Fixes issue #225.